+++ src/main/kotlin/org/uniplan/UniPlanApplication.kt
@@ -1,0 +1,11 @@
+package org.uniplan
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class UniPlanApplication
+
+fun main(args: Array<String>) {
+    runApplication<UniPlanApplication>(*args)
+}